refactor: complete EnhancePanel decomposition to 243 lines (#1030)
Full decomposition with state hook extraction:
- EnhanceSourcePreview.tsx (158 lines) — source waveform + A/B toggle
- EnhanceHistorySidebar.tsx (93 lines) — version tree + session list
- useEnhancePanelState.ts (463 lines) — all state, effects, callbacks

EnhancePanel.tsx: 1271 → 243 lines (81% reduction, under 300 target).
All 52 tests pass, zero type errors. All data-testid preserved.
